refactor: api 폴더로 파일 구조 정리 및 경로 수정 (#197)
