+++ src/features/auth/api/endpoints.ts
@@ -1,6 +1,6 @@
 export const API_ENDPOINTS = {
   SIGNUP: '/users/signup',
-  LOGIN: 'login',
-  LOGOUT: '/logout',
+  LOGIN: '/auth/login',
+  LOGOUT: '/auth/logout',
   REFRESH_TOKEN: '/token/refresh',
 };
