+++ src/features/auth/api/index.ts
@@ -1,3 +1,2 @@
 export * from './authAPI';
-export * from './backendAPI';
 export * from './endpoints';
